+++ app/api/animals/[id]/favorite/route.ts
@@ -28,6 +28,20 @@ export async function POST(req: NextRequest) {
   }
 
   try {
+
+    // 既に登録されているかを確認
+    const count = await prisma.individualAnimal.count({
+      where: {
+        animalId: Number(id),
+        individualId: body.userId,
+      },
+    });
+
+    if (count > 0) {
+      console.log("既にお気に入り登録済み。処理を終了します。");
+      return NextResponse.json({ message: "既にお気に入り登録済みです。" });
+    }
+
     // individualAnimal にお気に入り情報を登録
     const favorite = await prisma.individualAnimal.create({
       data: {
@@ -36,7 +50,11 @@ export async function POST(req: NextRequest) {
       },
     });
     console.log("お気に入り登録:", favorite);
-    return NextResponse.json(favorite);
+    // return NextResponse.json(favorite);
+    return NextResponse.json({
+      message: "お気に入り登録しました。",
+      isFavoriteProcessed: favorite ? true : false,
+    });
   } catch (error) {
     console.error("お気に入り登録エラー:", JSON.stringify(error));
     return NextResponse.json({ error: "お気に入り登録に失敗しました。" }, { status: 500 });
@@ -65,6 +83,20 @@ export async function DELETE(req: NextRequest) {
   }
 
   try {
+
+    // 既に登録されているかを確認
+    const count = await prisma.individualAnimal.count({
+      where: {
+        animalId: Number(id),
+        individualId: body.userId,
+      },
+    });
+
+    if (count === 0) {
+      console.log("お気に入り登録がありません。処理を終了します。");
+      return NextResponse.json({ message: "お気に入り登録がありません。" });
+    }
+
     // 複合ユニークキーを利用して、individualAnimal からお気に入り情報を削除
     const deletedFavorite = await prisma.individualAnimal.delete({
       where: {
@@ -75,7 +107,11 @@ export async function DELETE(req: NextRequest) {
       },
     });
     console.log("お気に入り解除:", deletedFavorite);
-    return NextResponse.json(deletedFavorite);
+
+    return NextResponse.json({
+      message: "お気に入り解除しました。",
+      isFavoriteProcessed: deletedFavorite ? true : false,
+    });
   } catch (error) {
     console.error("お気に入り解除エラー:", JSON.stringify(error));
     return NextResponse.json({ error: "お気に入り解除に失敗しました。" }, { status: 500 });
